+++ tools/releaser/test/fixtures/activestorage/lib/active_storage/gem_version.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
-module Rails
-  # Returns the currently loaded version of Rails as a <tt>Gem::Version</tt>.
+module ActiveStorage
+  # Returns the currently loaded version of \Rails as a +Gem::Version+.
   def self.gem_version
     Gem::Version.new VERSION::STRING
   end
 
   module VERSION
-    MAJOR = 7
+    MAJOR = 1
     MINOR = 0
-    TINY  = 8
-    PRE   = "4"
+    TINY  = 0
+    PRE   = "alpha"
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end