+++ actionview/CHANGELOG.md
@@ -1,3 +1,25 @@
+*   Include closing `</form>` tag when calling `form_tag` and `form_with` without a block:
+
+    ```ruby
+    config.action_view.closes_form_tag_without_block = true
+
+    form_tag "https://example.com"
+    # => <form action="https://example.com" method="post"><!-- Rails-generated hidden fields --></form>
+
+    form_with url: "https://example.com"
+    # => <form action="https://example.com" method="post"><!-- Rails-generated hidden fields --></form>
+
+    config.action_view.closes_form_tag_without_block = false
+
+    form_tag "https://example.com"
+    # => <form action="https://example.com" method="post"><!-- Rails-generated hidden fields -->
+
+    form_with url: "https://example.com"
+    # => <form action="https://example.com" method="post"><!-- Rails-generated hidden fields -->
+    ```
+
+    *Sean Doyle*
+
 ## Rails 8.0.0.beta1 (September 26, 2024) ##
 
 *   Enable DependencyTracker to evaluate renders with trailing interpolation.