+++ actionview/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Render `[accept-charset="UTF-8"]` on `<form>` elements rendered by
+    `button_to`
+
+    *Sean Doyle*
+
 ## Rails 8.0.0.beta1 (September 26, 2024) ##
 
 *   Enable DependencyTracker to evaluate renders with trailing interpolation.